Fix :Areggio commands to behave the same as :map
According to :help :command-bar, " and | in arguments for user-defined
Ex commands are eaten by the Vim script parser if the commands are
defined with -bar.

:Arpeggio and other Ex commands were defined with -bar automatically,
because most (= 80%, in 7.2.320) of built-in Ex commands are imaginarily
defined with -bar.

But :Arpeggio and others are variants of :map family, and :map family
does not accept either of " and |.  So that :Arpeggio and others should
behave the same as :map family.

Fix the behavior of :Arpeggio comands by removing -bar from their
definitions.

NB: :Arpeggio commands eventually execute :map family to define internal
key mappings, so that | given to :Arpeggio commands is eaten by the Vim
script parser.  This behavior is the same as one of :map family, but it
is different what :help :command-bar describes.